app/src/main/java/me/veryyoung/oj/leetcode/IntersectionOfTwoLinkedLists.java
package me.veryyoung.oj.leetcode;

import me.veryyoung.oj.util.ListNode;

/**
 * 160. Intersection of Two Linked Lists
 * https://leetcode.com/problems/intersection-of-two-linked-lists/
 */
public class IntersectionOfTwoLinkedLists {

    public ListNode getIntersectionNode(ListNode headA, ListNode headB) {
        int lengthA = getLength(headA);
        int lengthB = getLength(headB);
        if (lengthA > lengthB) {
            for (int i = 1; i <= lengthA - lengthB; i++) {
                headA = headA.next;
            }
        } else {
            for (int i = 1; i <= lengthB - lengthA; i++) {
                headB = headB.next;
            }
        }
        while (headA != headB) {
            headA = headA.next;
            headB = headB.next;
        }
        return headA;
    }

    private int getLength(ListNode head) {
        int length = 0;
        while (head != null) {
            length++;
            head = head.next;
        }
        return length;
    }
    
}
